+++ NEWS.md
@@ -3,7 +3,7 @@
 * Added covr and codecov
 * Removed extraneous color codes from log by adding stripping function 
 to `log_close()`.
-* Added path to source program to log header.
+* Added program path to log header.  Can now get this using `this.path()`.
 * Added `log_code()` function to dump code to the log.
 
 # logr 1.2.5